BE/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
pymysql
python-multipart
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose
alembic